Changes sala redencao regex strategy
Instead of attempting to encase the whole movie information in a single regex pattern,
we now try to match anything in the format \(Dir. Blabla | Country | 1999 | 120 min .
When a match is found we then run a separate regex for the screening dates.
This seems to accept a wider variety of possible formats